Change behavior of throws method
Previously it used to throw the value passed to the throws() call, like:
….with(1).throws(something). This 'something' had to be usually an
Error instance, but this turned out to be inconvenient. Since the Error
instance was created at the line of the .throws(something) call, it
didn't have a useful call stack from a place where it's thrown. Now this
has been changed. Any argument passed to the .throws() call will be
converted to a string and used as an error message for a newly created
Error instance.